Feature: Added settings UI to the app and weather request according to user preferred postal code
Added the required preferences, arrays and bools resource xml files for configuring properly a preference settings UI for the app according to the Android standard. Added the 'SettingsActivity.java' file, and added the settings activity to the manifest. Refactored the 'onOptionsItemSelected' method in the 'ForecastFragment.java' file to read the user's preferred location and pass it to the FetchWeather class.

Quiz: 3.09_Location from prefs

<Mid Lesson 3>